Restructure Python package and enable uvx execution from git repo
