+++ Python/README.md
@@ -4,31 +4,69 @@ Python bridge for interacting with Unreal Engine 5.5 using the Model Context Pro
 
 ## Setup
 
-1. Make sure Python 3.10+ is installed
-2. Install `uv` if you haven't already:
-   ```bash
-   curl -LsSf https://astral.sh/uv/install.sh | sh
-   ```
-3. Create and activate a virtual environment:
-   ```bash
-   uv venv
-   source .venv/bin/activate  # On Unix/macOS
-   # or
-   .venv\Scripts\activate     # On Windows
-   ```
-4. Install dependencies:
-   ```bash
-   uv pip install -e .
-   ```
+### 1. Install `uv` if you haven't already:
+```bash
+# On Unix/macOS:
+curl -LsSf https://astral.sh/uv/install.sh | sh
 
-At this point, you can configure your MCP Client (Claude Desktop, Cursor, Windsurf) to use the Unreal MCP Server as per the [Configuring your MCP Client](README.md#configuring-your-mcp-client).
+# On Windows (PowerShell):
+powershell -ExecutionPolicy ByPass -c "irm https://astral.sh/uv/install.ps1 | iex"
+```
+
+### 2. Configuring your MCP Client
+
+Use one of the following configurations in your MCP client's config file:
+
+### Option 1: Direct from Repository
+
+```json
+{
+  "mcpServers": {
+    "unrealMCP": {
+      "command": "uvx",
+      "args": [
+          "--from", 
+          "git+https://github.com/chongdashu/unreal-mcp#subdirectory=Python", 
+          "unreal_mcp_server"]
+    }
+  }
+}
+```
+
+#### Option 2: Local Installation
+
+```json
+{
+  "mcpServers": {
+    "unrealMCP": {
+      "command": "uv",
+      "args": [
+        "--directory", 
+        "<path/to/your/repo/Python>", 
+        "run", 
+        "unreal_mcp_server"
+      ]
+    }
+  }
+}
+```
+
+You can find the value of `<path/to/your/repo/>` by running `pwd` in the root directory of the repository.
+
+## Configuration Locations
+
+The location of your MCP configuration file depends on your MCP client. Common locations include:
+
+- Cursor: `~/.cursor/mcp_config.json`
+- Claude Desktop: `~/.config/claude/mcp_config.json`
+- Windsurf: `~/.codeium/windsurf/mcp_config.json`
+- Other clients: Refer to your client's documentation
 
 ## Testing Scripts
 
 There are several scripts in the [scripts](./scripts) folder. They are useful for testing the tools and the Unreal Bridge via a direct connection. This means that you do not need to have an MCP Server running.
 
 You should make sure you have installed dependencies and/or are running in the `uv` virtual environment in order for the scripts to work.
-
 
 ## Troubleshooting
