+++ Python/pyproject.toml
@@ -1,22 +1,23 @@
 [project]
 name = "unreal-mcp"
-version = "0.1.0"
+version = "0.2.0"
 description = "Unreal MCP Server: A server for Unreal Engine integration via the Model Context Protocol (MCP)."
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
   "mcp[cli]>=1.4.1",
   "fastmcp>=0.2.0",
-  "uvicorn",
-  "fastapi",
   "pydantic>=2.6.1",
   "requests"
 ]
+
+[project.scripts]
+unreal_mcp_server = "unreal_mcp.server:mcp.run"
 
 [build-system]
 requires = ["setuptools>=42", "wheel"]
 build-backend = "setuptools.build_meta"
 
-[tool.setuptools]
-# The main server script is a single-file module
-py-modules = ["unreal_mcp_server"] 
+[tool.setuptools.packages.find]
+where = ["."] # Explicitly look for packages in the current directory
+include = ["unreal_mcp*"]
