+++ Python/unreal_mcp/__init__.py
@@ -1,0 +1,1 @@
+# Unreal MCP package
