+++ Python/unreal_mcp/bridge/__init__.py
@@ -1,0 +1,1 @@
+# Unreal Bridge package
